backend/requirements.txt
fastapi==0.115.0
uvicorn==0.38.0
python-multipart
httpx==0.28.1
requests
PyMuPDF
pytesseract
Pillow
python-dotenv
